Add files via upload
implement a simple list, using singly linked-list data structure in C.
The data structure is a specified as follows:
A list contains zero or more elements. An element of a list can be either an atom
or a list. A list may also be empty, in which case is represented by NIL.
Two key operations are denied to access the elements of a list: car and cdr (see
Ref 6).
 Operation car (also: rst) takes a list as an argument and returns the head
of the list. The head of the list is the first element in the list.
 Operation cdr (also: rest) takes a list as an argument and returns the tail
of the list. The tail of the list, is a sub-list (list) that the head of the list is
pointing to. The tail of a empty list is considered empty.